Move scenes to scenes
