+++ src/components/index.js
@@ -1,3 +1,3 @@
 export { default as GlobalMenu } from './GlobalMenu'
-export { default as Home } from './Home'
+export { default as Header } from './Header'
 export { default as VersionWarning } from './VersionWarning'
