+++ src/scenes/Home/components/index.js
@@ -1,0 +1,3 @@
+export { Bold } from './Bold'
+export { Item } from './Item'
+export { Header } from './Header'
